DEV-23 added support to enable and disable logger

+++ src/Shared/StreamingHandlers/web_adapters.go
@@ -21,7 +21,7 @@ func newPacketWebAdapter(packet domain.Packet) PacketWebAdapter {
 		Name:         packet.Name,
 		Measurements: measurementWebAdapters,
 		Count:        packet.Count,
-		CycleTime:    packet.CycleTime,
+		CycleTime:    uint(packet.CycleTime),
 	}
 }
